requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
numpy
pydantic
pypdf